requirements.txt
pandas
cassandra-driver
jupyter
prettytable